Refactor and cleanup (more)
