+++ src/preload/moduleOverrides/generic-view.ts
@@ -1,6 +1,6 @@
 import replaceTopLevelDelete from "./partials/replaceTopLevelDelete";
 import replaceDisplayIndex from "./partials/replaceDisplayIndex";
-import { DependencyNameMap } from "preload/withDependencyMap";
+import { DependencyNameMap } from "../overrideHelpers/withDependencyMap";
 
 export default (dependencyNameMap: DependencyNameMap) => ({
   /* Applies to image-view, table-view, and text_view */
