+++ src/preload/script.ts
@@ -3,7 +3,7 @@ import moduleOverrides from "./moduleOverrides";
 import { postMessageUp, listenToMessageDown } from "utils/messages";
 import injectScript from "utils/injectScript";
 import { pollForValue } from "utils/utils";
-import withDependencyMap from "./withDependencyMap";
+import withDependencyMap from "./overrideHelpers/withDependencyMap";
 
 /* This script is loaded at document_start, before the page's scripts, to give it 
 time to set ALMOND_OVERRIDES and replace module definitions */
